add metnoom to IPAM plan
We currently are backing up netmoom to "metnoom" the new DS923+ with
20.7TB, which is going to serve as the secondary (and then primary) DNS,
while we reformat and blow out the dust in Netmoom.

It is online now with 10.17.13.140 assigned by DHCP, but it should not
have a DHCP-assigned address as it will be serving DHCP, DNS, and others
soon. It has been added to the plan at 10.17.13.250, an unused address
in the 13-net subnet range in the neighborhood of Netmoom's 13-net IP.

Signed-off-by: Kingdon B <kingdon@urmanac.com>